+++ ChangeLog.markdown
@@ -7,6 +7,7 @@ __N/A__
 
 * The convert transformation can no longer be triggered via the t query parameter per default
 * Fixed #150: Error model can't be formatted using an image formatter
+* Fixed #149: Add mode to the border transformation
 
 Imbo-0.1.0
 ----------